refactor(e2e): split large test files for better CI parallelization
